Implement better microsecond interrupt system
Handles multiple interrupts scheduled at the same time.
Little bit less code.

+++ src/microint.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 
 typedef struct {
-    void (*function)(uint16_t);
-    uint16_t code;
+    void (*function)(int);
+    int code;
     uint16_t tick;
 } MicroInt;
